refactor(zig-runtime): continue f64 to f32 narrowing across runtime paths (#278)
* refactor(replay): narrow relative movement and heading path to f32

* refactor(spawn): narrow internal survival and rush scalar math to f32

* refactor(weapons): narrow runtime scalar math and angle helpers to f32

* refactor(replay): narrow nominal elapsed summary math to f32

* refactor(creatures): narrow helper RNG scalar generation to f32

* refactor(creatures): narrow phase-seed generation sites to f32

* test(runtime): align perks and bonuses float assertions with f32 state

* test(creatures): narrow spawn fixture expectations to f32

* refactor(projectiles): narrow half-pi and test float helper to f32

* refactor(creatures): narrow terrain and shot world-size boundaries to f32

* refactor(creatures): narrow spawn heading helper path to f32

* refactor(spawn): narrow tint clamp path to f32 internals

* test(spawn): narrow float expectation fixtures to f32

* refactor(creatures): infer spawn-init widening casts without explicit f64

* refactor(spawn): infer float widening in stage and tint helper paths

* refactor(spawn): inline generic appendSpawnCall float casts

* refactor(spawn): align stage2 milestone spacing with native f32 step

* refactor(spawn): align python stage2 milestone spacing to native f32

* refactor(spawn): narrow rush and survival paths to f32 cadence

* refactor(spawn): narrow quest heading payloads to f32

* refactor(spawn): narrow quest geometry vectors to f32

* refactor(runtime-math): remove unused native helper exports

* test(quest-spawn): replace summary hash with row assertions

* refactor(replay): narrow capture game-state floats to f32

* refactor(replay): switch wire float schema to f32

* refactor(replay): narrow wire integer schema to 32-bit

* refactor(zig): remove redundant f32 casts in spawn paths

+++ crimson-zig/src/quest_spawn/dump_main.zig
@@ -43,8 +43,8 @@ pub fn main() !void {
         try writeStderr("invalid seed\n");
         std.process.exit(1);
     };
-    const world_size: f64 = if (args.len == 6)
-        std.fmt.parseFloat(f64, args[5]) catch {
+    const world_size: f32 = if (args.len == 6)
+        std.fmt.parseFloat(f32, args[5]) catch {
             try writeStderr("invalid world_size\n");
             std.process.exit(1);
         }
@@ -97,7 +97,7 @@ fn buildByImpl(
     level_key: i32,
     player_count: i32,
     seed: u32,
-    world_size: f64,
+    world_size: f32,
     out_entries: []spawn_runtime.QuestSpawnEntry,
 ) !quest_spawn_logic_full.QuestSpawnBuildResult {
     if (std.mem.eql(u8, impl, "main")) {
